Flatten Storage packages.
