+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/masks/MaskedStorageInferred.java
@@ -11,7 +11,7 @@ import org.enso.table.problems.BlackholeProblemAggregator;
 final class MaskedStorageInferred<T> extends MaskedStorage<T>
     implements ColumnStorageWithInferredStorage {
   private ColumnStorage<?> cachedInferredStorage = null;
-  private boolean hasSpecializedStorageBeenInferred = false;
+  private boolean hasComputedInferredStorage = false;
 
   MaskedStorageInferred(ColumnStorage<T> parent, IndexMapper indexMapper) {
     super(parent, indexMapper);
@@ -19,7 +19,7 @@ final class MaskedStorageInferred<T> extends MaskedStorage<T>
 
   @Override
   public ColumnStorage<?> getInferredStorage() {
-    if (!hasSpecializedStorageBeenInferred) {
+    if (!hasComputedInferredStorage) {
       var inferredType = CastOperation.reconcileObjectStorage(this);
       cachedInferredStorage =
           (inferredType instanceof AnyObjectType)
@@ -29,7 +29,7 @@ final class MaskedStorageInferred<T> extends MaskedStorage<T>
                   true,
                   Builder.getForType(inferredType, getSize(), BlackholeProblemAggregator.INSTANCE),
                   (builder, index, value) -> builder.append(value));
-      hasSpecializedStorageBeenInferred = true;
+      hasComputedInferredStorage = true;
     }
 
     return cachedInferredStorage;
