+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/AbstractLongStorage.java
@@ -1,8 +1,5 @@
-package org.enso.table.data.column.storage.numeric;
+package org.enso.table.data.column.storage;
 
-import org.enso.table.data.column.storage.ColumnLongStorage;
-import org.enso.table.data.column.storage.Storage;
-import org.enso.table.data.column.storage.ValueIsNothingException;
 import org.enso.table.data.column.storage.iterators.ColumnLongStorageIterator;
 import org.enso.table.data.column.storage.iterators.LongStorageIterator;
 import org.enso.table.data.column.storage.type.IntegerType;
