+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/ComputedLongStorage.java
@@ -1,8 +1,7 @@
-package org.enso.table.data.column.storage.numeric;
+package org.enso.table.data.column.storage;
 
 import java.util.BitSet;
-import org.enso.table.data.column.storage.ColumnStorageWithNothingMap;
-import org.enso.table.data.column.storage.ValueIsNothingException;
+
 import org.enso.table.data.column.storage.type.IntegerType;
 
 /**
