+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/DoubleStorage.java
@@ -1,13 +1,9 @@
-package org.enso.table.data.column.storage.numeric;
+package org.enso.table.data.column.storage;
 
 import java.util.BitSet;
 import java.util.NoSuchElementException;
-import org.enso.table.data.column.storage.ColumnDoubleStorage;
-import org.enso.table.data.column.storage.ColumnStorageWithNothingMap;
-import org.enso.table.data.column.storage.Storage;
-import org.enso.table.data.column.storage.ValueIsNothingException;
+
 import org.enso.table.data.column.storage.iterators.ColumnDoubleStorageIterator;
-import org.enso.table.data.column.storage.iterators.DoubleStorageIterator;
 import org.enso.table.data.column.storage.type.FloatType;
 
 /** A column containing floating point numbers. */
