+++ std-bits/table/src/main/java/org/enso/table/operations/AddRowNumber.java
@@ -1,7 +1,7 @@
 package org.enso.table.operations;
 
 import org.enso.table.data.column.storage.ColumnStorage;
-import org.enso.table.data.column.storage.numeric.LongStorage;
+import org.enso.table.data.column.storage.LongStorage;
 import org.enso.table.data.column.storage.type.IntegerType;
 import org.enso.table.data.table.Column;
 import org.enso.table.problems.ProblemAggregator;
